feat(evidence): #116 WS2 — portable browser-safe verify-core
Factor the §9.2 verification logic out of verify.ts into a pure,
browser-safe verify-core/ module so civicaitools.org (server) and the
future typedstandards.org browser verifier (WS3) verify off ONE source
and cannot drift. Node→browser swaps: @noble/hashes for SHA-256/512
(was node:crypto), a fixed-prefix SPKI DER slice for the raw Ed25519 key
(was crypto.createPublicKey→JWK), atob/Uint8Array/TextDecoder (no
Buffer), and injected FetchLike for all network I/O (Rekor, blob refs).
RFC 8785 JCS and the @noble/curves Ed25519/Ed25519ph verify are reused
as-is, including the #111 algorithm dispatch.

- New verify-core/ (types, primitives, canonicalization, profiles,
  attestation consts, blob-ref, signature, trust-registry, checks,
  lifecycle, rekor, verifyEvidence orchestrator, index barrel).
- The server modules (verify.ts, canonicalization.ts, profiles.ts,
  blob-ref.ts, signing.ts, attestation.ts) now re-export FROM verify-core;
  verify.ts keeps only the server-only trust-registry loader (fs/process).
- The verify route consumes verifyEvidence; output is byte-identical
  (server-deeper lifecycle + live rekor preserved). All existing verify
  tests (#110/#111) + the full suite still pass (283 tests).
- ESLint no-restricted-imports guard (scoped to verify-core/**) forbids
  node:crypto/fs/path/process so a future Node import fails CI.
- @noble/hashes promoted to a direct dependency.

Check depth matches the current server: presence-only #7, hash-parity #8,
lifecycle STATE #10 in the portable path. Deeper offline crypto + the
authoritative Q15 bundle test are #119. WS2 makes verify-core exist,
portable, server-consumed, and parity-proven; it does not build the
verifier UI (WS3) or harden the crypto (#119).
