+++ package.json
@@ -18,6 +18,7 @@
   "dependencies": {
     "@neondatabase/serverless": "^1.0.2",
     "@noble/curves": "^2.2.0",
+    "@noble/hashes": "^2.2.0",
     "@vercel/blob": "^2.3.3",
     "@vercel/kv": "^3.0.0",
     "@vercel/sandbox": "^1.10.2",
